+++ mtdata/main.py
@@ -24,12 +24,13 @@ def list_data(langs, names, not_names=None, full=False):
     print(f"Total {len(entries)} entries")
 
 
-def get_data(langs, out_dir, train_dids=None, test_dids=None, dev_did=None, merge_train=False, **kwargs):
+def get_data(langs, out_dir, train_dids=None, test_dids=None, dev_did=None, merge_train=False, compress=False,
+             **kwargs):
     from mtdata.data import Dataset
     assert train_dids or test_dids, 'Required --train or --test or both'
     dataset = Dataset.prepare(
         langs, train_dids=train_dids, test_dids=test_dids, out_dir=out_dir,
-        dev_did=dev_did, cache_dir=CACHE_DIR, merge_train=merge_train)
+        dev_did=dev_did, cache_dir=CACHE_DIR, merge_train=merge_train, compress=compress)
     cli_sig = f'-l {"-".join(str(l) for l in langs)}'
     if train_dids:
         cli_sig += f' -tr {" ".join(str(d) for d in train_dids)}'
@@ -95,20 +96,6 @@ def lang_pair(string) -> LangPair:
     return std_codes
 
 
-def dataset_id(string) -> DatasetId:
-    expected_format = "<group>-<name>-<version>-<l1>-<l2>"
-    parts = string.strip().split('-')
-    if len(parts) != 5:
-        raise argparse.ArgumentTypeError(f'Dataset ID expected in format: {expected_format}; but given {string}.'
-                                         f' If you are unsure, run "mtdata list | grep -i <name>" and copy its id.')
-    group, name, version, lang1, lang2 = parts
-    langs = lang_pair(f'{lang1}-{lang2}')
-    try:
-        did = DatasetId(group=group, name=name, version=version, langs=langs)
-    except Exception as e:
-        raise argparse.ArgumentTypeError(e)
-    return did
-
 
 def add_boolean_arg(parser: argparse.ArgumentParser, name, dest=None, default=False, help=''):
     group = parser.add_mutually_exclusive_group()
@@ -149,19 +136,20 @@ def parse_args():
     get_p.add_argument('-l', '--langs', metavar='L1-L2', type=lang_pair,
                        help='Language pairs; e.g.: deu-eng',
                        required=True)
-    get_p.add_argument('-tr', '--train', metavar='ID', dest='train_dids', nargs='*', type=dataset_id,
+    get_p.add_argument('-tr', '--train', metavar='ID', dest='train_dids', nargs='*', type=DatasetId.parse,
                        help='''R|Names of datasets separated by space, to be used for *training*.
     e.g. -tr Statmt-news_commentary-16-deu-eng europarl_v9 .
      To concatenate all these into a single train file, set --merge flag.''')
-    get_p.add_argument('-ts', '--test', metavar='ID', dest='test_dids', nargs='*', type=dataset_id,
+    get_p.add_argument('-ts', '--test', metavar='ID', dest='test_dids', nargs='*', type=DatasetId.parse,
                        help='''R|Names of datasets separated by space, to be used for *testing*. 
     e.g. "-ts Statmt-newstest_deen-2019-deu-eng Statmt-newstest_deen-2020-deu-eng ".
     You may also use shell expansion if your shell supports it.
     e.g. "-ts Statmt-newstest_deen-20{19,20}-deu-eng" ''')
-    get_p.add_argument('-dev', '--dev', metavar='ID', dest='dev_did', type=dataset_id, required=False,
+    get_p.add_argument('-dev', '--dev', metavar='ID', dest='dev_did', type=DatasetId.parse, required=False,
                        help='''R|Dataset to be used for development (aka validation). 
     e.g. "-dev Statmt-newstest_deen-2017-deu-eng"''')
     add_boolean_arg(get_p, 'merge', dest='merge_train', default=False, help='Merge train into a single file')
+    get_p.add_argument(f'--compress', action='store_true', default=False, help="Keep the files compressed")
 
     get_p.add_argument('-o', '--out', dest='out_dir', type=Path, required=True, help='Output directory name')